src/mcp_server/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
python-dotenv
chromadb
sentence-transformers
mcp